# tools/dashboard/backend/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/polardbxclusters/parameter-templates:
    get:
      consumes:
      - application/json
      description: Lists all PolarDB-X parameter templates in the specified namespace
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of parameter templates
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: List parameter templates
      tags:
      - parameter-templates
    post:
      consumes:
      - application/json
      description: Creates a new PolarDB-X parameter template
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      - description: Parameter template specification
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created parameter template
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameter template specification
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Parameter template already exists
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Create parameter template
      tags:
      - parameter-templates
  /api/v1/polardbxclusters/parameter-templates/{namespace}/{name}:
    delete:
      consumes:
      - application/json
      description: Deletes a PolarDB-X parameter template
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the parameter template
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Parameter template not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Delete parameter template
      tags:
      - parameter-templates
    get:
      consumes:
      - application/json
      description: Retrieves details of a specific PolarDB-X parameter template
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the parameter template
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Parameter template details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Parameter template not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Get parameter template
      tags:
      - parameter-templates
    put:
      consumes:
      - application/json
      description: Updates an existing PolarDB-X parameter template
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the parameter template
        in: path
        name: name
        required: true
        type: string
      - description: Updated parameter template specification
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated parameter template
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameter template specification
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Parameter template not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Update parameter template
      tags:
      - parameter-templates
  /api/v1/polardbxclusters/parameters:
    get:
      consumes:
      - application/json
      description: Lists all PolarDB-X parameters in the specified namespace
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of parameters
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: List parameters
      tags:
      - parameters
    post:
      consumes:
      - application/json
      description: Creates a new PolarDB-X parameter
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      - description: Parameter specification
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created parameter
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameter specification
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Parameter already exists
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Create parameter
      tags:
      - parameters
  /api/v1/polardbxclusters/parameters/{name}:
    delete:
      consumes:
      - application/json
      description: Deletes a PolarDB-X parameter
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      - description: Name of the parameter
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Parameter not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Delete parameter
      tags:
      - parameters
    get:
      consumes:
      - application/json
      description: Retrieves details of a specific PolarDB-X parameter
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      - description: Name of the parameter
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Parameter details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Parameter not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Get parameter
      tags:
      - parameters
    put:
      consumes:
      - application/json
      description: Updates an existing PolarDB-X parameter
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      - description: Name of the parameter
        in: path
        name: name
        required: true
        type: string
      - description: Updated parameter specification
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated parameter
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameter specification
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Parameter not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Update parameter
      tags:
      - parameters
...
swagger: "2.0"